tweak MultiSelect demo page CSS and improve section event handling Type Safety
